scraper::element_ref

Struct ElementRef

pub struct ElementRef<'a> { /* private fields */ }

Wrapper around a reference to an element node.

This wrapper implements the Element trait from the selectors crate, which allows it to be matched against CSS selectors.

Implementations

impl<'a> ElementRef<'a>

pub fn wrap(node: NodeRef<'a, Node>) -> Option<Self>

Wraps a NodeRef only if it references a Node::Element.

pub fn value(&self) -> &'a Element

Returns the Element referenced by self.

pub fn select<'b>(&self, selector: &'b Selector) -> Select<'a, 'b>

Returns an iterator over descendent elements matching a selector.

pub fn html(&self) -> String

Returns the HTML of this element.

pub fn inner_html(&self) -> String

Returns the inner HTML of this element.

pub fn attr(&self, attr: &str) -> Option<&'a str>

Returns the value of an attribute.

pub fn text(&self) -> Text<'a>

Returns an iterator over descendent text nodes.

pub fn child_elements(&self) -> impl Iterator<Item = ElementRef<'a>>

Iterate over all child nodes which are elements

Example

let fragment = Html::parse_fragment("foo<span>bar</span><a>baz</a>qux");

let children = fragment.root_element().child_elements().map(|element| element.value().name()).collect::<Vec<_>>();
assert_eq!(children, ["span", "a"]);

pub fn descendent_elements(&self) -> impl Iterator<Item = ElementRef<'a>>

Iterate over all descendent nodes which are elements

Example

let fragment = Html::parse_fragment("foo<span><b>bar</b></span><a><i>baz</i></a>qux");

let descendants = fragment.root_element().descendent_elements().map(|element| element.value().name()).collect::<Vec<_>>();
assert_eq!(descendants, ["html", "span", "b", "a", "i"]);