scraper

Module selector

CSS selectors.

Structs

CssLocalName

Wraps LocalName so that it can be used with selectors

CssString

Wraps String so that it can be used with selectors

Parser

An implementation of Parser for selectors

Selector

Wrapper around CSS selectors.

Simple

A simple implementation of SelectorImpl with no pseudo-classes or pseudo-elements.

Enums

NonTSPseudoClass

Non Tree-Structural Pseudo-Class.

PseudoElement

CSS Pseudo-Element

Traits

ToCss

Trait for things the can serialize themselves in CSS syntax.