Struct scraper::element_ref::ElementRef

pub struct ElementRef<'a> { /* fields omitted */ }

Wrapper around a reference to an element node.

This wrapper implements the Element trait from the selectors crate, which allows it to be matched against CSS selectors.

Methods

impl<'a> ElementRef<'a>

pub fn wrap(node: NodeRef<'a, Node>) -> Option<Self>

Wraps a NodeRef only if it references a Node::Element.

pub fn value(&self) -> &'a Element

Returns the Element referenced by self.

Important traits for Select<'a, 'b>
impl<'a, 'b> Iterator for Select<'a, 'b> type Item = ElementRef<'a>;

pub fn select<'b>(&self, selector: &'b Selector) -> Select<'a, 'b>

Returns an iterator over descendent elements matching a selector.

pub fn html(&self) -> String

Returns the HTML of this element.

pub fn inner_html(&self) -> String

Returns the inner HTML of this element.

Important traits for Text<'a>
impl<'a> Iterator for Text<'a> type Item = &'a str;

pub fn text(&self) -> Text<'a>

Returns an iterator over descendent text nodes.