Struct scraper::html::Html

pub struct Html {
    pub errors: Vec<Cow<'static, str>>,
    pub quirks_mode: QuirksMode,
    pub tree: Tree<Node>,
}

An HTML tree.

Parsing does not fail hard. Instead, the quirks_mode is set and errors are added to the errors field. The tree will still be populated as best as possible.

Implements the TreeSink trait from the html5ever crate, which allows HTML to be parsed.

Fields
errors: Vec<Cow<'static, str>>

Parse errors.

quirks_mode: QuirksMode

The quirks mode.

tree: Tree<Node>

The node tree.

Methods

impl Html

pub fn new_document() -> Self

Creates an empty HTML document.

pub fn new_fragment() -> Self

Creates an empty HTML fragment.

pub fn parse_document(document: &str) -> Self

Parses a string of HTML as a document.

pub fn parse_fragment(fragment: &str) -> Self

Parses a string of HTML as a fragment.

Important traits for Select<'a, 'b>
impl<'a, 'b> Iterator for Select<'a, 'b> type Item = ElementRef<'a>;

pub fn select<'a, 'b>(&'a self, selector: &'b Selector) -> Select<'a, 'b>

Returns an iterator over elements matching a selector.