Struct scraper::html::Html

pub struct Html {
    pub errors: Vec<Cow<'static, str>>,
    pub quirks_mode: QuirksMode,
    pub tree: Tree<Node>,
}

An HTML tree.

Parsing does not fail hard. Instead, the quirks_mode is set and errors are added to the errors field. The tree will still be populated as best as possible.

Implements the TreeSink trait from the html5ever crate, which allows HTML to be parsed.

Fields

errors

Parse errors.

quirks_mode

The quirks mode.

tree

The node tree.

Methods

impl Html

fn new_document() -> Self

Creates an empty HTML document.

fn new_fragment() -> Self

Creates an empty HTML fragment.

fn parse_document(document: &str) -> Self

Parses a string of HTML as a document.

fn parse_fragment(fragment: &str) -> Self

Parses a string of HTML as a fragment.

fn select<'a, 'b>(&'a self, selector: &'b Selector) -> Select<'a, 'b>

Returns an iterator over elements matching a selector.